Deduplicate the `vulnerability` spec
