+++ cyclonedx-bom/src/specs/common/mod.rs
@@ -32,6 +32,7 @@ pub(crate) mod property;
 pub(crate) mod service;
 pub(crate) mod signature;
 pub(crate) mod tool;
+pub(crate) mod vulnerability;
 pub(crate) mod vulnerability_analysis;
 pub(crate) mod vulnerability_credits;
 pub(crate) mod vulnerability_rating;
