+++ cyclonedx-bom/src/specs/v1_4/mod.rs
@@ -18,11 +18,10 @@
 
 pub(crate) mod bom;
 
-pub(crate) mod vulnerability;
-
 pub(crate) use crate::specs::common::component::v1_4 as component;
 pub(crate) use crate::specs::common::composition::v1_4 as composition;
 pub(crate) use crate::specs::common::external_reference::v1_4 as external_reference;
 pub(crate) use crate::specs::common::metadata::v1_4 as metadata;
 pub(crate) use crate::specs::common::service::v1_4 as service;
 pub(crate) use crate::specs::common::tool::v1_4 as tool;
+pub(crate) use crate::specs::common::vulnerability::v1_4 as vulnerability;
