+++ cyclonedx-bom/src/specs/v1_5/mod.rs
@@ -19,7 +19,6 @@
 pub(crate) mod annotation;
 pub(crate) mod bom;
 pub(crate) mod lifecycles;
-pub(crate) mod vulnerability;
 
 pub(crate) use crate::specs::common::component::v1_5 as component;
 pub(crate) use crate::specs::common::composition::v1_5 as composition;
@@ -28,3 +27,4 @@ pub(crate) use crate::specs::common::metadata::v1_5 as metadata;
 #[allow(unused_imports)]
 pub(crate) use crate::specs::common::service::v1_5 as service;
 pub(crate) use crate::specs::common::tool::v1_5 as tool;
+pub(crate) use crate::specs::common::vulnerability::v1_5 as vulnerability;
